% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_continuous.R
\name{ab_continuous}
\alias{ab_continuous}
\title{A super-simple continuous-time two-compartment model}
\usage{
ab_continuous(rate = 0.1, time_points = seq(0, 10, by = 0.1))
}
\arguments{
\item{rate}{the rate of progression from A to B per unit time (must be positive)}

\item{time_points}{a vector of time points to include in the data frame returned}
}
\description{
A super-simple continuous-time two-compartment model
}
