% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_discrete.R
\name{ab_discrete}
\alias{ab_discrete}
\title{A super-simple discrete-time two-compartment model}
\usage{
ab_discrete(rate = 0.1, d_time = 0.1, max_time = 10)
}
\arguments{
\item{rate}{the rate of progression from A to B per unit time (must be positive)}

\item{d_time}{the desired time step (delta time)}

\item{max_time}{the desired maximum time point (must be greater than the time step)}
}
\description{
A super-simple discrete-time two-compartment model
}
