% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_methods.R
\name{autoplot.ipdmr_ct}
\alias{autoplot.ipdmr_ct}
\title{Autoplot methods associated with IPDMR model functions}
\usage{
\method{autoplot}{ipdmr_ct}(object, ...)
}
\arguments{
\item{object}{an object returned by an IPDMR model function}

\item{...}{additional arguments (currently ignored)}
}
\description{
Autoplot methods associated with IPDMR model functions
}
