% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{group_models}
\alias{group_models}
\alias{sir_det}
\alias{seir_det}
\alias{sir_stoc}
\alias{seir_stoc}
\title{Discrete-time deterministic and stochastic SIR and SEIR models}
\usage{
sir_det(
  S = 99,
  I = 1,
  R = 0,
  beta = 0.25,
  gamma = 0.2,
  delta = 0.05,
  transmission_type = c("frequency", "density"),
  d_time = 1,
  max_time = 100
)

seir_det(
  S = 99,
  E = 0,
  I = 1,
  R = 0,
  numE = 3L,
  beta = 0.25,
  omega = 0.2,
  gamma = 0.2,
  delta = 0.05,
  vacc = 0,
  repl = 0,
  cull = 0,
  transmission_type = c("frequency", "density"),
  d_time = 1,
  max_time = 100
)

sir_stoc(
  S = 99,
  I = 1,
  R = 0,
  beta = 0.25,
  gamma = 0.2,
  delta = 0.05,
  transmission_type = c("frequency", "density"),
  d_time = 1,
  max_time = 100,
  iterations = 1L
)

seir_stoc(
  S = 99,
  E = 0,
  I = 1,
  R = 0,
  numE = 3L,
  beta = 0.25,
  omega = 0.2,
  gamma = 0.2,
  delta = 0.05,
  vacc = 0,
  repl = 0,
  cull = 0,
  transmission_type = c("frequency", "density"),
  d_time = 1,
  max_time = 100,
  iterations = 1L
)
}
\arguments{
\item{S}{the starting number of susceptible animals}

\item{I}{the starting number of infected animals}

\item{R}{the starting number of recovered animals}

\item{beta}{the transmission rate parameter per unit time (must be positive)}

\item{gamma}{the recovery rate parameter per unit time (must be positive)}

\item{delta}{the reversion rate parameter per unit time (must be positive)}

\item{transmission_type}{either frequency or density}

\item{d_time}{the desired time step (delta time)}

\item{max_time}{the desired maximum time point (must be greater than the time step)}

\item{E}{the total starting number of exposed/latent animals (this will be distributed among the sub-compartments of E when the model is initialised)}

\item{numE}{the number of sub-compartments within the exposed/latent state}

\item{omega}{the latent progression rate parameter per unit time (must be positive)}

\item{vacc}{the vaccination rate parameter per unit time (must be positive)}

\item{repl}{the replacement rate parameter per unit time (must be positive)}

\item{cull}{the targeted culling rate parameter per unit time (must be positive)}

\item{iterations}{the number of iterations to run (stochastic models only)}
}
\description{
These wrapper functions create a single-group model, run the model for the
indicated time steps, and then return a data frame representing the model
output. Note that starting values for S/E/I/R can be continuous or discrete
for the deterministic models, but must be discrete numbers for the stochastic
models.
}
