% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group.R
\name{make_group}
\alias{make_group}
\title{Title}
\usage{
make_group(
  update_type = c("deterministic", "stochastic"),
  numE = 3L,
  d_time = 0.1,
  group_name = NA_character_,
  model_type = c("SEIR", "SIR", "SI"),
  implementation = c("R", "C++")
)
}
\arguments{
\item{implementation}{}
}
\description{
Title
}
