% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_wrappers.R
\name{multi_models}
\alias{multi_models}
\alias{multi_seir_det}
\alias{multi_seir_stoc}
\title{Discrete-time deterministic and stochastic SEIR models for multiple groups}
\usage{
multi_seir_det(
  n_groups,
  beta_matrix,
  S = 99,
  E = 0,
  I = 1,
  R = 0,
  numE = 3L,
  beta = 0.25,
  omega = 0.2,
  gamma = 0.2,
  delta = 0.05,
  vacc = 0,
  repl = 0,
  cull = 0,
  d_time = 1,
  max_time = 100
)

multi_seir_stoc(
  n_groups,
  beta_matrix,
  S = 99,
  E = 0,
  I = 1,
  R = 0,
  numE = 3L,
  beta = 0.25,
  omega = 0.2,
  gamma = 0.2,
  delta = 0.05,
  vacc = 0,
  repl = 0,
  cull = 0,
  d_time = 1,
  max_time = 100,
  iterations = 1L
)
}
\arguments{
\item{n_groups}{the number of groups to generate}

\item{beta_matrix}{a matrix representing between-group spread (must have nrow and ncol equal to n_groups)}

\item{S}{the starting number of susceptible animals}

\item{E}{the total starting number of exposed/latent animals (this will be distributed among the sub-compartments of E when the model is initialised)}

\item{I}{the starting number of infected animals}

\item{R}{the starting number of recovered animals}

\item{numE}{the number of sub-compartments within the exposed/latent state}

\item{beta}{the transmission rate parameter per unit time (must be positive)}

\item{omega}{the latent progression rate parameter per unit time (must be positive)}

\item{gamma}{the recovery rate parameter per unit time (must be positive)}

\item{delta}{the reversion rate parameter per unit time (must be positive)}

\item{vacc}{the vaccination rate parameter per unit time (must be positive)}

\item{repl}{the replacement rate parameter per unit time (must be positive)}

\item{cull}{the targeted culling rate parameter per unit time (must be positive)}

\item{d_time}{the desired time step (delta time)}

\item{max_time}{the desired maximum time point (must be greater than the time step)}

\item{iterations}{the number of iterations to run (stochastic models only)}
}
\description{
These wrapper functions create a multi-group model, run the model for the
indicated time steps, and then return a data frame representing the model
output. Note that starting values for S/E/I/R can be continuous or discrete
for the deterministic models, but must be discrete numbers for the stochastic
models.
}
\examples{
output <- multi_seir_det(2, matrix(0,nrow=2,ncol=2))
ggplot2::autoplot(output)

output <- multi_seir_stoc(2, matrix(0,nrow=2,ncol=2), iterations=3L)
ggplot2::autoplot(output)
}
