% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_continuous.R
\name{si_continuous}
\alias{si_continuous}
\title{Title}
\usage{
si_continuous(
  N = 10,
  beta = 0.05,
  type = c("frequency", "density"),
  init_I = 1,
  time_points = seq(0, 21, by = 0.1)
)
}
\arguments{
\item{time_points}{}
}
\description{
Title
}
\examples{
si_continuous(N=10, type="density") |> ggplot2::autoplot()
si_continuous(N=10, type="frequency") |> ggplot2::autoplot()

}
