% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_discrete.R
\name{si_discrete}
\alias{si_discrete}
\title{Title}
\usage{
si_discrete(
  N = 10,
  beta = 0.05,
  type = c("frequency", "density"),
  init_I = 1,
  time_step = 1/24,
  max_time = 21
)
}
\arguments{
\item{init_I}{}
}
\description{
Title
}
\examples{
si_discrete(N=10, type="density") |> ggplot2::autoplot()
si_discrete(N=10, type="frequency") |> ggplot2::autoplot()

}
