% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{sir_det}
\alias{sir_det}
\title{Title}
\usage{
sir_det(
  S = 99,
  I = 1,
  R = 0,
  beta = 0.25,
  gamma = 0.2,
  delta = 0.05,
  transmission_type = "frequency",
  d_time = 1L,
  max_time = 100L
)
}
\arguments{
\item{max_time}{}
}
\description{
Title
}
