% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_wrappers.R
\name{sirs_det}
\alias{sirs_det}
\title{Title}
\usage{
sirs_det(
  S = 99,
  I = 1,
  R = 0,
  beta = 0.2,
  gamma = 0.2,
  delta = 0.2,
  transmission_type = "frequency",
  time_step = 1L,
  max_time = 100L
)
}
\arguments{
\item{max_time}{}
}
\description{
Title
}
