% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithinGroupModel.R
\name{WithinGroupModel}
\alias{WithinGroupModel}
\title{WithinGroupModel}
\description{
General within-group model class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WithinGroupModel-new}{\code{WithinGroupModel$new()}}
\item \href{#method-WithinGroupModel-update}{\code{WithinGroupModel$update()}}
\item \href{#method-WithinGroupModel-check_state}{\code{WithinGroupModel$check_state()}}
\item \href{#method-WithinGroupModel-run}{\code{WithinGroupModel$run()}}
\item \href{#method-WithinGroupModel-clone}{\code{WithinGroupModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WithinGroupModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-WithinGroupModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WithinGroupModel$new(
  model_type = c("sirs"),
  update_type = c("deterministic", "stochastic"),
  transmission_type = c("frequency", "density"),
  time_step = 1L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WithinGroupModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-WithinGroupModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WithinGroupModel$update(time_step = self$time_step)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WithinGroupModel-check_state"></a>}}
\if{latex}{\out{\hypertarget{method-WithinGroupModel-check_state}{}}}
\subsection{Method \code{check_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WithinGroupModel$check_state()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WithinGroupModel-run"></a>}}
\if{latex}{\out{\hypertarget{method-WithinGroupModel-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WithinGroupModel$run(
  n_steps,
  time_step = self$time_step,
  include_current = self$time == 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WithinGroupModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WithinGroupModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WithinGroupModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
