% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rates.R
\name{apply_rates}
\alias{apply_rates}
\title{Title}
\usage{
apply_rates(
  compartments,
  rates,
  d_time,
  update_type = c("deterministic", "stochastic")
)
}
\arguments{
\item{update_type}{deterministic (default) or stochastic}
}
\description{
Title
}
