% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenGroupModel.R
\name{BetweenGroupModel}
\alias{BetweenGroupModel}
\title{BetweenGroupModel}
\description{
A general-purpose between-group model spread class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BetweenGroupModel-new}{\code{BetweenGroupModel$new()}}
\item \href{#method-BetweenGroupModel-update}{\code{BetweenGroupModel$update()}}
\item \href{#method-BetweenGroupModel-save}{\code{BetweenGroupModel$save()}}
\item \href{#method-BetweenGroupModel-reset}{\code{BetweenGroupModel$reset()}}
\item \href{#method-BetweenGroupModel-run}{\code{BetweenGroupModel$run()}}
\item \href{#method-BetweenGroupModel-clone}{\code{BetweenGroupModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$new(models)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-update"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$update(d_time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-save"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$save()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-reset"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-run"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$run(n_steps, d_time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BetweenGroupModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BetweenGroupModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetweenGroupModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
