% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPDMR-package.R
\docType{package}
\name{IPDMR-package}
\alias{IPDMR-package}
\alias{IPDMR}
\title{Support Code for the Introduction to Practical Disease Modelling Course}
\description{
A collection of functions and classes that are used to illustrate fundamental concepts of disease modelling as part of a physical course series taught jointly by the University of Copenhagen and University of Sydney. The package is not intended to be used for other purposes; you are welcome to do so, but please note that the package may change without notice in order to suit changes to the associated taught course. All functions, arguments and classes are therefore subject to change without following any form of deprecation process.
}
\details{
A collection of functions and classes that are used to illustrate
fundamental concepts of disease modelling as part of a physical course series
taught jointly by the University of Copenhagen and University of Sydney. The
package is not intended to be used for other purposes; you are welcome to
do so, but please note that the package may change without notice in order
to suit changes to the associated taught course. All functions, arguments and
classes are therefore subject to change without following any form of
deprecation process.
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Other contributors:
\itemize{
  \item Carsten Kirkeby \email{ckir@sund.ku.dk} [reviewer]
  \item Michael Ward \email{michael.ward@sydney.edu.au} [reviewer]
}

}
\keyword{internal}
