% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group.R
\name{make_group}
\alias{make_group}
\title{Generate an object representing a within-group spread model}
\usage{
make_group(
  update_type = c("deterministic", "stochastic"),
  numE = 3L,
  numI = 1L,
  numR = 1L,
  group_name = NA_character_,
  implementation = c("C++", "R6")
)
}
\arguments{
\item{update_type}{either stochastic or deterministic}

\item{numE}{the number of sub-compartments desired for the E state (0 or more)}

\item{numI}{the number of sub-compartments desired for the I state (1 or more)}

\item{numR}{the number of sub-compartments desired for the R state (0 or more)}

\item{group_name}{an optional name for the group}

\item{implementation}{either C++ or R6}
}
\description{
This is a wrapper function around the new/initialise methods for R6 and C++
implementations of within-group spread models used on the IPDMR course.
For more information on available methods see the documentation for
\link{SEIRclass} (the method signatures and outputs of the R6 and C++
implementations should be identical).
}
\examples{
set.seed(2024-11-05)
r6res <- make_group(update_type = "stochastic", implementation = "R6")$run(10, 0.1)

set.seed(2024-11-05)
cppres <- make_group(update_type = "stochastic", implementation = "C++")$run(10, 0.1)

## Should be identical:
stopifnot(unlist(r6res) - unlist(cppres) == 0L)

}
