% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_boxes.R
\name{suggest_boxes}
\alias{suggest_boxes}
\title{Suggest a number of sub-compartments to obtain a suitable distribution of transition times}
\usage{
suggest_boxes(mean, sd)
}
\arguments{
\item{mean}{the desired mean waiting time}

\item{sd}{the desired standard deviation in waiting time}
}
\value{
the function returns a ggplot object illustrating the suggestions,
and prints the suggested number of sub-compartments and rates to screen
}
\description{
This is a utility function that takes a desired mean and standard deviation for
the distribution of transit times through a disease compartment, and suggests
a suitable rate and number of sub-compartments that matches the provided mean
and variance as closely as possible.
}
\examples{
# An average of 5 days, with sd of 2.5 days:
suggest_boxes(mean=5, sd=3)

# The number of sub-compartments is invariant to transformation of the rate
# on different time scales, i.e. we get the same if specifying in hours:
suggest_boxes(mean=5*24, sd=3*24)

}
